requirements.txt
aiogram==3.26.0
python-dotenv==1.0.1
pytest==8.3.3



